+++ packages/xstate-solid/src/createService.ts
@@ -1,81 +1,11 @@
-import type {
-  AnyStateMachine,
-  InterpreterOptions,
-  Observer,
-  AreAllImplementationsAssumedToBeProvided,
-  InternalMachineOptions,
-  InterpreterFrom
-} from 'xstate';
-import { State, interpret, toObserver } from 'xstate';
-import type { UseMachineOptions } from './types';
-import { onCleanup, onMount } from 'solid-js';
-
-type RestParams<
-  TMachine extends AnyStateMachine
-> = AreAllImplementationsAssumedToBeProvided<
-  TMachine['__TResolvedTypesMeta']
-> extends false
-  ? [
-      options: InterpreterOptions &
-        UseMachineOptions<TMachine['__TContext'], TMachine['__TEvent']> &
-        InternalMachineOptions<
-          TMachine['__TContext'],
-          TMachine['__TEvent'],
-          TMachine['__TResolvedTypesMeta'],
-          true
-        >,
-      observerOrListener?:
-        | Observer<
-            State<
-              TMachine['__TContext'],
-              TMachine['__TEvent'],
-              any,
-              TMachine['__TTypestate'],
-              TMachine['__TResolvedTypesMeta']
-            >
-          >
-        | ((
-            value: State<
-              TMachine['__TContext'],
-              TMachine['__TEvent'],
-              any,
-              TMachine['__TTypestate'],
-              TMachine['__TResolvedTypesMeta']
-            >
-          ) => void)
-    ]
-  : [
-      options?: InterpreterOptions &
-        UseMachineOptions<TMachine['__TContext'], TMachine['__TEvent']> &
-        InternalMachineOptions<
-          TMachine['__TContext'],
-          TMachine['__TEvent'],
-          TMachine['__TResolvedTypesMeta']
-        >,
-      observerOrListener?:
-        | Observer<
-            State<
-              TMachine['__TContext'],
-              TMachine['__TEvent'],
-              any,
-              TMachine['__TTypestate'],
-              TMachine['__TResolvedTypesMeta']
-            >
-          >
-        | ((
-            value: State<
-              TMachine['__TContext'],
-              TMachine['__TEvent'],
-              any,
-              TMachine['__TTypestate'],
-              TMachine['__TResolvedTypesMeta']
-            >
-          ) => void)
-    ];
+import type { AnyStateMachine, InterpreterFrom } from 'xstate';
+import { State, interpret } from 'xstate';
+import type { RestParams } from './types';
+import { onCleanup } from 'solid-js';
 
 export function createService<TMachine extends AnyStateMachine>(
   machine: TMachine,
-  ...[options = {}, observerOrListener]: RestParams<TMachine>
+  ...[options = {}]: RestParams<TMachine>
 ): InterpreterFrom<TMachine> {
   const {
     context,
@@ -105,17 +35,8 @@ export function createService<TMachine extends AnyStateMachine>(
     rehydratedState ? (State.create(rehydratedState) as any) : undefined
   );
 
-  onMount(() => {
-    let sub;
-
-    if (observerOrListener) {
-      sub = service.subscribe(toObserver(observerOrListener));
-    }
-
-    onCleanup(() => {
-      service.stop();
-      sub?.unsubscribe();
-    });
+  onCleanup(() => {
+    service.stop();
   });
 
   return service as InterpreterFrom<TMachine>;
