+++ packages/xstate-solid/src/fsm.ts
@@ -77,9 +77,7 @@ export function useService<TService extends StateMachine.AnyService>(
         checkReusedService(getServiceState(service()));
       }
       const { unsubscribe } = service().subscribe((nextState) => {
-        setState(
-          reconcile<typeof nextState, typeof nextState>(nextState)
-        );
+        setState(reconcile<typeof nextState, typeof nextState>(nextState));
       });
       onCleanup(unsubscribe);
     })
