+++ packages/xstate-solid/src/types.ts
@@ -1,9 +1,14 @@
 import type { EventObject, StateConfig } from 'xstate';
+import {
+  AnyStateMachine,
+  AreAllImplementationsAssumedToBeProvided,
+  InternalMachineOptions,
+  InterpreterFrom,
+  InterpreterOptions,
+  StateFrom
+} from 'xstate';
 
-export type MaybeLazy<T> = T | (() => T);
-export type Prop<T, K> = K extends keyof T ? T[K] : never;
-
-export interface UseMachineOptions<TContext, TEvent extends EventObject> {
+interface UseMachineOptions<TContext, TEvent extends EventObject> {
   /**
    * If provided, will be merged with machine's `context`.
    */
@@ -14,3 +19,35 @@ export interface UseMachineOptions<TContext, TEvent extends EventObject> {
    */
   state?: StateConfig<TContext, TEvent>;
 }
+
+export type UseMachineReturn<
+  TMachine extends AnyStateMachine,
+  TInterpreter = InterpreterFrom<TMachine>
+> = [StateFrom<TMachine>, Prop<TInterpreter, 'send'>, TInterpreter];
+
+export type Prop<T, K> = K extends keyof T ? T[K] : never;
+
+export type RestParams<
+  TMachine extends AnyStateMachine
+> = AreAllImplementationsAssumedToBeProvided<
+  TMachine['__TResolvedTypesMeta']
+> extends false
+  ? [
+      options: InterpreterOptions &
+        UseMachineOptions<TMachine['__TContext'], TMachine['__TEvent']> &
+        InternalMachineOptions<
+          TMachine['__TContext'],
+          TMachine['__TEvent'],
+          TMachine['__TResolvedTypesMeta'],
+          true
+        >
+    ]
+  : [
+      options?: InterpreterOptions &
+        UseMachineOptions<TMachine['__TContext'], TMachine['__TEvent']> &
+        InternalMachineOptions<
+          TMachine['__TContext'],
+          TMachine['__TEvent'],
+          TMachine['__TResolvedTypesMeta']
+        >
+    ];
